formatting and tweaks
