+++ src/coreclr/jit/fgdiagnostic.cpp
@@ -11,7 +11,6 @@
 
 // Flowgraph Check and Dump Support
 
-
 #ifdef DEBUG
 void Compiler::fgPrintEdgeWeights()
 {
