+++ src/coreclr/jit/fgopt.cpp
@@ -7,10 +7,9 @@
 #pragma hdrstop
 #endif
 
-#include "lower.h"       // for LowerRange()
+#include "lower.h" // for LowerRange()
 
 // Flowgraph Optimization
-
 
 /*****************************************************************************
  *
