+++ src/coreclr/jit/fgstmt.cpp
@@ -9,7 +9,6 @@
 
 // Flowgraph Statements
 
-
 #ifdef DEBUG
 // Check to see if block contains a statement but don't spend more than a certain
 // budget doing this per method compiled.
